// frontend/src/company.d.ts
export interface CompanyCashFlow {
  accessNumber: string;
  symbol: string;
  year: number;
  quarter: number;
  form: string;
  startDate: string;
  endDate: string;
  report: {
    cf: {
      concept: string;
      label: string;
      unit: string;
      value: number;
    }[];
  };
}